move isObject to utils
